// src/app/api/models/media-asset-response-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface MediaAssetResponseDto {
  contentUrl: string;
  creator?: string;
  height: number;
  id: string;
  license: 'user_owned' | 'cc0' | 'pdm' | 'legacy_unknown';
  licenseUrl?: string;
  publishable: boolean;
  source: 'upload' | 'openverse' | 'legacy';
  sourceUrl?: string;
  thumbnailUrl: string;
  title?: string;
  width: number;
}