// src/app/api/models/media-search-result-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface MediaSearchResultDto {
  attribution: string;
  creator?: string;
  id: string;
  license: 'cc0' | 'pdm';
  licenseUrl?: string;
  source?: string;
  sourceUrl: string;
  thumbnailUrl: string;
  title: string;
}